% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpeciesModels.R
\name{pseudo_absence_naive}
\alias{pseudo_absence_naive}
\title{Pseudo Absences (Naive method)
Given two sets: relative_sample and background, returns 0, 1 or NA.
This is necessary for defining absences in presence-only data.}
\usage{
pseudo_absence_naive(relative_sample, background)
}
\description{
Pseudo Absences (Naive method)
Given two sets: relative_sample and background, returns 0, 1 or NA.
This is necessary for defining absences in presence-only data.
}
\examples{
pseudo_absence_naive()
}
